Fix input text overlay
Previously, clicking on the default text in the form
would not activate input. The input was placed in front of
the text, so it is activated by the click. But, the text
could not be seen. Setting background to transparent
gives the same visual, while allowing input to be activated.